+++ sdk/textanalytics/Azure.AI.TextAnalytics/src/PiiEntityDomain.cs
@@ -6,7 +6,7 @@ namespace Azure.AI.TextAnalytics
     /// <summary>
     /// The different domains of PII entities that users can filter requests by.
     /// </summary>
-    public enum PiiEntityDomainType
+    public enum PiiEntityDomain
     {
         /// <summary>
         /// Don't apply any domain filter. This is the default value.
@@ -20,14 +20,14 @@ namespace Azure.AI.TextAnalytics
     }
 
     [System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1649:File name should match first type name", Justification = "Small extensions, good to keep here.")]
-    internal static class PiiEntityDomainTypeExtensions
+    internal static class PiiEntityDomainExtensions
     {
-        internal static string GetString(this PiiEntityDomainType type)
+        internal static string GetString(this PiiEntityDomain type)
         {
             return type switch
             {
-                PiiEntityDomainType.None => null,
-                PiiEntityDomainType.ProtectedHealthInformation => "phi",
+                PiiEntityDomain.None => null,
+                PiiEntityDomain.ProtectedHealthInformation => "phi",
                 _ => null,
             };
         }